refactor (just toss some code around to be intuitively findable)
