+++ xtask/src/commands/mod.rs
@@ -1,241 +1,47 @@
-use std::{
-    collections::HashMap,
-    path::{Path, PathBuf},
-};
-
-use anyhow::{Context, Result, bail};
-use clap::Args;
+use anyhow::{Result, bail};
 use inquire::Select;
-use serde::Serialize;
-use strum::IntoEnumIterator;
 
-pub use self::{build::*, check_changelog::*, check_pr_changelog::*, release::*, run::*};
-use crate::{
-    Package,
-    cargo::{CargoAction, CargoCommandBatcher},
-    firmware,
-    metadata::Chip,
+pub use self::{
+    build::*,
+    check::*,
+    check_changelog::*,
+    check_pr_changelog::*,
+    ci::*,
+    clean::*,
+    dispatch::DispatchArgs,
+    fmt::*,
+    host_tests::*,
+    lint::*,
+    release::*,
+    run::*,
 };
+use crate::{cargo::CargoAction, metadata::Chip};
+
 mod build;
+mod check;
 mod check_changelog;
 mod check_pr_changelog;
+mod ci;
+mod clean;
 pub mod dispatch;
+mod examples;
+mod fmt;
 #[cfg(feature = "report")]
 pub mod generate_report;
 #[cfg(feature = "semver-checks")]
 pub(crate) mod generate_rom_symbols;
+mod host_tests;
+mod lint;
 #[cfg(feature = "mcp")]
 pub mod mcp;
 #[cfg(feature = "rel-check")]
 pub mod relcheck;
 mod release;
 mod run;
+mod tests;
 
-// ----------------------------------------------------------------------------
-// Top level command and subcommand definitions
-
-/// Arguments for the `ci` subcommand.
-#[cfg_attr(
-    feature = "mcp",
-    xtask_mcp_macros::mcp_tool(
-        description = "Perform (parts of) the checks done in CI for a given chip",
-        command = "ci"
-    )
-)]
-#[derive(Debug, Args)]
-pub struct CiArgs {
-    /// Chip to target.
-    #[arg(value_enum)]
-    pub chip: Chip,
-
-    /// The toolchain used to run the lints
-    #[arg(long)]
-    pub toolchain: Option<String>,
-
-    /// Steps to run in the CI pipeline.
-    #[arg(long, value_delimiter = ',')]
-    pub steps: Vec<String>,
-
-    /// Whether to skip running lints
-    #[arg(long)]
-    pub no_lint: bool,
-
-    /// Whether to skip building documentation
-    #[arg(long)]
-    pub no_docs: bool,
-
-    /// Whether to skip checking the crates itself
-    #[arg(long)]
-    pub no_check_crates: bool,
-}
-
-/// Arguments for the `fmt` subcommand.
-#[cfg_attr(
-    feature = "mcp",
-    xtask_mcp_macros::mcp_tool(
-        description = "Format all packages in the workspace with rustfmt",
-        command = "fmt"
-    )
-)]
-#[derive(Debug, Args)]
-pub struct FmtPackagesArgs {
-    /// Run in 'check' mode; exits with 0 if formatted correctly, 1 otherwise
-    #[arg(long)]
-    pub check: bool,
-
-    /// Package(s) to target.
-    #[arg(value_enum, default_values_t = Package::iter())]
-    pub packages: Vec<Package>,
-}
-
-/// Arguments for the `clean` subcommand.
-#[cfg_attr(
-    feature = "mcp",
-    xtask_mcp_macros::mcp_tool(
-        description = "Run cargo clean for the specified packages",
-        command = "clean"
-    )
-)]
-#[derive(Debug, Args)]
-pub struct CleanArgs {
-    /// Package(s) to target.
-    #[arg(value_enum, default_values_t = Package::iter())]
-    pub packages: Vec<Package>,
-}
-
-/// Arguments for the `host-tests` subcommand.
-#[cfg_attr(
-    feature = "mcp",
-    xtask_mcp_macros::mcp_tool(
-        description = "Run host-tests in the workspace with cargo test",
-        command = "host-tests"
-    )
-)]
-#[derive(Debug, Args)]
-pub struct HostTestsArgs {
-    /// Package(s) to target.
-    #[arg(value_enum, default_values_t = Package::iter())]
-    pub packages: Vec<Package>,
-}
-
-/// Arguments for checking packages (`check` without an example/test name, and
-/// `check-global-symbols`).
-#[derive(Debug, Args)]
-pub struct CheckPackagesArgs {
-    /// Package(s) to target.
-    #[arg(value_enum, default_values_t = Package::iter())]
-    pub packages: Vec<Package>,
-
-    /// Check for a specific chip
-    #[arg(long, value_enum, value_delimiter = ',', default_values_t = Chip::iter())]
-    pub chips: Vec<Chip>,
-
-    /// The toolchain used to run the checks
-    #[arg(long)]
-    pub toolchain: Option<String>,
-}
-
-/// Arguments for the `lint` subcommand.
-#[cfg_attr(
-    feature = "mcp",
-    xtask_mcp_macros::mcp_tool(
-        description = "Lint all packages in the workspace with clippy",
-        command = "lint"
-    )
-)]
-#[derive(Debug, Args)]
-pub struct LintPackagesArgs {
-    /// Package(s) to target.
-    #[arg(value_enum, default_values_t = Package::iter())]
-    pub packages: Vec<Package>,
-
-    /// Lint for a specific chip
-    #[arg(long, value_enum, value_delimiter = ',', default_values_t = Chip::iter())]
-    pub chips: Vec<Chip>,
-
-    /// Automatically apply fixes
-    #[arg(long)]
-    pub fix: bool,
-
-    /// The toolchain used to run the lints
-    #[arg(long)]
-    pub toolchain: Option<String>,
-}
-
-/// Flags and free-form tokens shared by `build` / `run` / `check` / `test`.
-#[cfg_attr(
-    feature = "mcp",
-    xtask_mcp_macros::mcp_tool(verbs(
-        build = "Build an example, crate, or tests. Tokens: chip, crate, example, test, or package alias (examples, tests, qa), in any order.",
-        run = "Flash and run an example or qa binary. Tokens: chip, example name, or package alias (examples, qa). Not for HIL tests — use `test`. Chip is inferred from a connected device when omitted.",
-        check = "Check crates with cargo check, or try-build examples and tests. Tokens: chip, crate, example, test, or package alias. No tokens checks all published crates.",
-        test = "Run HIL tests only. Tokens: chip and/or test name. Not for examples — use `run`. Chip is inferred from probe-rs when omitted.",
-    ))
-)]
-#[derive(Debug, Args)]
-pub struct DispatchArgs {
-    /// Chip(s) to target. Inferred from a connected device when omitted.
-    #[arg(long, value_enum, value_delimiter = ',')]
-    pub chip: Vec<Chip>,
-
-    /// Package(s) to act on.
-    #[arg(long, value_enum, value_delimiter = ',')]
-    pub package: Vec<Package>,
-
-    /// Build examples in debug mode.
-    #[arg(long)]
-    pub debug: bool,
-
-    /// Toolchain used to build.
-    #[arg(long)]
-    pub toolchain: Option<String>,
-
-    /// Emit crate build timings.
-    #[arg(long)]
-    pub timings: bool,
-
-    /// Repeat HIL tests this many times.
-    #[arg(long, default_value_t = 1)]
-    pub repeat: usize,
-
-    /// HIL test selector(s). Also accepted as free-form tokens.
-    #[arg(long, short = 't', alias = "tests", value_delimiter = ',', num_args = 1..)]
-    pub test: Option<Vec<String>>,
-
-    /// Chip, crate, example, or test names, in any order.
-    pub tokens: Vec<String>,
-}
-
-// ----------------------------------------------------------------------------
-// Subcommand Arguments
-
-fn standalone_example_projects(package: Package) -> bool {
-    matches!(package, Package::Examples | Package::CompileTests)
-}
-
-/// Arguments common to commands which act on doctests.
-#[cfg_attr(
-    feature = "mcp",
-    xtask_mcp_macros::mcp_tool(
-        description = "Run doc tests for the specified chip and packages",
-        command = "doc-tests"
-    )
-)]
-#[derive(Debug, Args)]
-pub struct DocTestArgs {
-    /// Package(s) where we wish to run doc tests.
-    #[arg(long, value_enum, value_delimiter = ',', default_values_t = Package::iter())]
-    pub packages: Vec<Package>,
-    /// Chip to target.
-    #[arg(value_enum)]
-    pub chip: Chip,
-}
-
-// ----------------------------------------------------------------------------
-// Subcommand Actions
-
-const EXAMPLE_ARGUMENT_HINT: &str =
-    "the example name as the first argument, or `all` to act on every example of the package";
+pub use examples::examples;
+pub use tests::tests;
 
 /// Asks the caller to pick one of `options`.
 ///
@@ -254,555 +60,7 @@ pub(crate) fn select<T: std::fmt::Display>(
     Ok(Select::new(message, options).prompt()?)
 }
 
-/// Execute the given action on the specified examples.
-pub fn examples(
-    workspace: &Path,
-    package: Package,
-    chip: Chip,
-    example: Option<&str>,
-    action: CargoAction,
-    debug: bool,
-    toolchain: Option<&str>,
-    timings: bool,
-) -> Result<()> {
-    log::debug!("Running examples for '{package}' on '{chip:?}'");
-
-    package
-        .validate_package_chip(&chip)
-        .with_context(|| format!("The package '{package}' does not support the chip '{chip:?}'"))?;
-
-    let package_path = crate::windows_safe_path(&workspace.join(package.directory()));
-
-    // Load all examples which support the specified chip and parse their metadata.
-    //
-    // Directories might contain a number of individual projects, and don't not rely on
-    // metadata comments in the source files. As such, it needs to load its metadata differently
-    // than other packages.
-    let examples = if standalone_example_projects(package) {
-        crate::firmware::load_cargo_toml(&package_path).with_context(|| {
-            format!(
-                "Failed to load specified examples from {}",
-                package_path.display()
-            )
-        })?
-    } else {
-        let example_path = match package {
-            Package::QaTest => package_path.join("src").join("bin"),
-            _ => package_path.join("examples"),
-        };
-
-        crate::firmware::load(&example_path)?
-    };
-
-    let mut examples = examples
-        .into_iter()
-        .filter(|example| example.supports_chip(chip))
-        .collect::<Vec<_>>();
-
-    examples.sort_by_key(|a| a.binary_name());
-
-    let mut filtered = vec![];
-
-    if let Some(example) = example {
-        filtered.clone_from(&examples);
-        if !example.eq_ignore_ascii_case("all") {
-            filtered.retain(|ex| ex.matches_name(example));
-
-            if filtered.is_empty() {
-                log::warn!(
-                    "Example '{example}' not found or unsupported for the given chip. Please select one of the existing examples in the desired package."
-                );
-
-                let example_name = select(
-                    "Select the example:",
-                    examples
-                        .iter()
-                        .map(|ex| ex.lookup_name(&package_path))
-                        .collect(),
-                    EXAMPLE_ARGUMENT_HINT,
-                )?;
-
-                if let Some(selected) = examples.iter().find(|ex| ex.matches_name(&example_name)) {
-                    filtered.push(selected.clone());
-                }
-            }
-        }
-    } else {
-        let example_name = select(
-            "Select an example:",
-            examples
-                .iter()
-                .map(|ex| ex.lookup_name(&package_path))
-                .collect(),
-            EXAMPLE_ARGUMENT_HINT,
-        )?;
-
-        if let Some(selected) = examples.iter().find(|ex| ex.matches_name(&example_name)) {
-            filtered.push(selected.clone());
-        }
-    }
-
-    match action {
-        CargoAction::Build(out_path) => build_examples(
-            package,
-            chip,
-            debug,
-            toolchain,
-            timings,
-            filtered,
-            &package_path,
-            out_path.as_deref(),
-        ),
-        CargoAction::Run => run_examples(
-            package,
-            chip,
-            debug,
-            toolchain,
-            timings,
-            filtered,
-            &package_path,
-        ),
-    }
-}
-
-/// Execute the given action on the specified HIL tests.
-pub fn tests(
-    workspace: &Path,
-    package: Package,
-    chip: Chip,
-    test: Option<&[String]>,
-    action: CargoAction,
-    repeat: usize,
-    toolchain: Option<&str>,
-    timings: bool,
-) -> Result<()> {
-    if !matches!(package, Package::HilTest | Package::HilTestRadio) {
-        bail!("Unknown test package: {package}. Use 'hil-test' or 'hil-test-radio'");
-    }
-
-    let package_path = crate::windows_safe_path(&workspace.join(package.directory()));
-    let is_radio_package = package == Package::HilTestRadio;
-
-    // Determine the appropriate build target for the given package and chip:
-    let target = Package::HilTest.target_triple(&chip)?;
-
-    // Load all test metadata first so we can differentiate between:
-    // - unknown test selectors (typos)
-    // - valid tests that are unsupported for the selected chip.
-    let bins_root = package_path.join("src").join("bin");
-    let mut all_tests = crate::firmware::load(&bins_root)?;
-    if is_radio_package {
-        let tests_root = bins_root.join("tests");
-        if tests_root.exists() {
-            all_tests.extend(crate::firmware::load(&tests_root)?);
-        }
-        let support_root = bins_root.join("support");
-        if support_root.exists() {
-            all_tests.extend(crate::firmware::load(&support_root)?);
-        }
-    }
-    all_tests.sort_by_key(|a| a.binary_name());
-    let tests_for_chip = all_tests
-        .iter()
-        .filter(|example| example.supports_chip(chip))
-        .collect::<Vec<_>>();
-
-    // Radio tests reuse the normal HIL flow; the only difference is a pre-step
-    // that builds and flashes the support firmware on a secondary probe.
-    if let CargoAction::Build(Some(out_dir)) = &action {
-        // Make sure the tmp directory has no garbage for us.
-        let tmp_dir = out_dir.join("tmp");
-        _ = std::fs::remove_dir_all(&tmp_dir);
-        std::fs::create_dir_all(&tmp_dir).unwrap();
-    }
-
-    let mut commands = CargoCommandBatcher::new();
-    let mut artifact_meta = HashMap::<String, firmware::Metadata>::new();
-    // Execute the specified action:
-    // If user sets some specific test(s)
-    if let Some(test_arg) = test {
-        let trimmed: Vec<&str> = test_arg.iter().map(|s| s.trim()).collect();
-
-        // Reject `--test ""` / `--test " "`
-        if trimmed.iter().all(|s| s.is_empty()) {
-            bail!("Empty test selector is not allowed.");
-        }
-
-        let mut unknown_selected: Vec<String> = Vec::new();
-        let mut unsupported_for_chip: Vec<String> = Vec::new();
-
-        for selected in trimmed.into_iter().filter(|s| !s.is_empty()) {
-            let (test_arg, filter) = match selected.split_once("::") {
-                Some((test, filter)) => (Some(test), Some(filter)),
-                None => (Some(selected), None),
-            };
-
-            let run_test_extra_args = if is_radio_package && action == CargoAction::Run {
-                &[]
-            } else {
-                (action == CargoAction::Run)
-                    .then(|| filter.as_ref().map(std::slice::from_ref))
-                    .flatten()
-                    .unwrap_or(&[])
-            };
-
-            let matched: Vec<_> = all_tests.iter().filter(|t| t.matches(test_arg)).collect();
-
-            if matched.is_empty() {
-                if is_radio_package
-                    && let Some(radio_meta) = resolve_radio_module_alias(&all_tests, chip, selected)
-                {
-                    let command = crate::generate_build_command(
-                        &package_path,
-                        chip,
-                        &target,
-                        &radio_meta,
-                        action.clone(),
-                        false,
-                        toolchain,
-                        timings,
-                        &[],
-                    )?;
-                    artifact_meta.insert(command.artifact_name.clone(), radio_meta);
-                    commands.push(command);
-                    continue;
-                }
-                unknown_selected.push(selected.to_string());
-            } else {
-                let matched_for_chip: Vec<_> = matched
-                    .into_iter()
-                    .filter(|t| t.supports_chip(chip))
-                    .collect();
-
-                if matched_for_chip.is_empty() {
-                    unsupported_for_chip.push(selected.to_string());
-                    continue;
-                }
-
-                for test in matched_for_chip {
-                    let command = crate::generate_build_command(
-                        &package_path,
-                        chip,
-                        &target,
-                        test,
-                        action.clone(),
-                        false,
-                        toolchain,
-                        timings,
-                        run_test_extra_args,
-                    )?;
-                    artifact_meta.insert(command.artifact_name.clone(), test.clone());
-                    commands.push(command);
-                }
-            }
-        }
-
-        if !unsupported_for_chip.is_empty() {
-            log::warn!(
-                "Skipping unsupported tests for '{}': {}",
-                chip,
-                unsupported_for_chip.join(", ")
-            );
-        }
-
-        if !unknown_selected.is_empty() {
-            bail!("Test not found: {}", unknown_selected.join(", "))
-        }
-    } else {
-        for test in tests_for_chip {
-            let command = crate::generate_build_command(
-                &package_path,
-                chip,
-                &target,
-                test,
-                action.clone(),
-                false,
-                toolchain,
-                timings,
-                &[],
-            )?;
-            artifact_meta.insert(command.artifact_name.clone(), test.clone());
-            commands.push(command);
-        }
-    }
-
-    let mut harness_for_artifact = HashMap::<String, firmware::Metadata>::new();
-    if is_radio_package {
-        for test in artifact_meta.values() {
-            if is_support_firmware(test) {
-                continue;
-            }
-            let Some(harness_name) = test.harness_firmware() else {
-                continue;
-            };
-
-            let Some(harness_meta) = all_tests.iter().find(|candidate| {
-                if !candidate.supports_chip(chip) {
-                    return false;
-                }
-
-                let harness_basename = harness_name.rsplit('/').next().unwrap_or(harness_name);
-                candidate.binary_name() == harness_name
-                    || candidate.output_file_name() == harness_name
-                    || candidate.binary_name() == harness_basename
-                    || candidate.output_file_name() == harness_basename
-            }) else {
-                bail!(
-                    "Missing harness firmware '{harness_name}' required by '{}'",
-                    test.output_file_name()
-                );
-            };
-
-            harness_for_artifact.insert(test.output_file_name(), harness_meta.clone());
-        }
-    }
-
-    // Ensure required harness firmware is built so the artifact is
-    // self-contained for `run elfs` on the HIL runner. Not needed on Run:
-    // `build_radio_harness` builds harnesses lazily per test.
-    if is_radio_package && matches!(action, CargoAction::Build(_)) {
-        use std::collections::HashSet;
-
-        let mut queued: HashSet<String> = artifact_meta
-            .values()
-            .map(firmware::Metadata::output_file_name)
-            .collect();
-
-        let mut harnesses_to_add: Vec<firmware::Metadata> = Vec::new();
-        for harness in harness_for_artifact.values() {
-            if queued.insert(harness.output_file_name()) {
-                harnesses_to_add.push(harness.clone());
-            }
-        }
-
-        for harness in harnesses_to_add {
-            let command = crate::generate_build_command(
-                &package_path,
-                chip,
-                &target,
-                &harness,
-                action.clone(),
-                false,
-                toolchain,
-                timings,
-                &[],
-            )?;
-            artifact_meta.insert(command.artifact_name.clone(), harness);
-            commands.push(command);
-        }
-    }
-
-    let mut failed = Vec::new();
-    let built_commands = commands.build(false);
-    let built_artifacts = artifact_meta.keys().cloned().collect::<Vec<_>>();
-
-    let mut built_harness = HashMap::<String, PathBuf>::new();
-
-    for c in built_commands {
-        let repeat = if matches!(action, CargoAction::Run) {
-            repeat
-        } else {
-            1
-        };
-
-        let radio_meta = artifact_meta.get(&c.artifact_name);
-
-        // Support firmware never runs as a DUT; it's used only as a pre-step
-        // for whichever DUT requests it via `HARNESS-FIRMWARE`.
-        if matches!(action, CargoAction::Run) && radio_meta.is_some_and(is_support_firmware) {
-            log::info!("Skipping support firmware '{}' as DUT", c.artifact_name);
-            continue;
-        }
-
-        println!(
-            "Command: cargo {}",
-            c.command.join(" ").replace("---", "\n    ---")
-        );
-
-        for i in 0..repeat {
-            if repeat != 1 {
-                log::info!("Run {}/{}", i + 1, repeat);
-            }
-
-            if matches!(action, CargoAction::Run)
-                && let Some(harness_meta) = harness_for_artifact.get(&c.artifact_name)
-            {
-                let harness_path = match build_radio_harness(
-                    workspace,
-                    &package_path,
-                    chip,
-                    &target,
-                    toolchain,
-                    timings,
-                    harness_meta,
-                    &mut built_harness,
-                ) {
-                    Ok(path) => path,
-                    Err(e) => {
-                        failed.push(c.artifact_name.clone());
-                        log::error!(
-                            "Failed to build harness firmware '{}' for '{}': {e}",
-                            harness_meta.output_file_name(),
-                            c.artifact_name
-                        );
-                        break;
-                    }
-                };
-
-                let harness_spec = crate::radio_hil_runner::HarnessSpec {
-                    binary_path: &harness_path,
-                    target: &target,
-                };
-                if let Err(e) =
-                    crate::radio_hil_runner::run_radio_test(&c, Some(harness_spec), None)
-                {
-                    failed.push(c.artifact_name.clone());
-                    log::error!("Radio test '{}' failed: {}", c.artifact_name, e);
-                    break;
-                }
-            } else if c.run(false).is_err() {
-                failed.push(c.artifact_name.clone());
-            }
-        }
-    }
-
-    move_artifacts(chip, &action);
-
-    if !failed.is_empty() {
-        bail!("Failed tests: {:#?}", failed);
-    }
-
-    if is_radio_package && let CargoAction::Build(Some(out_dir)) = &action {
-        write_radio_elf_manifest(
-            out_dir,
-            chip,
-            &built_artifacts,
-            &artifact_meta,
-            &harness_for_artifact,
-        )?;
-    }
-
-    Ok(())
-}
-
-const RADIO_ELF_MANIFEST_FILENAME: &str = "radio-elfs.json";
-
-#[derive(Debug, Serialize)]
-struct RadioElfManifestEntry {
-    artifact: String,
-    harness: Option<String>,
-    support_firmware: bool,
-}
-
-fn write_radio_elf_manifest(
-    out_dir: &Path,
-    chip: Chip,
-    built_artifacts: &[String],
-    artifact_meta: &HashMap<String, firmware::Metadata>,
-    harness_for_artifact: &HashMap<String, firmware::Metadata>,
-) -> Result<()> {
-    let mut entries = Vec::new();
-    for artifact in built_artifacts {
-        let Some(meta) = artifact_meta.get(artifact) else {
-            continue;
-        };
-        entries.push(RadioElfManifestEntry {
-            artifact: artifact.clone(),
-            harness: harness_for_artifact
-                .get(artifact)
-                .map(firmware::Metadata::output_file_name),
-            support_firmware: is_support_firmware(meta),
-        });
-    }
-
-    if entries.is_empty() {
-        return Ok(());
-    }
-
-    let manifest_path = out_dir
-        .join(chip.to_string())
-        .join(RADIO_ELF_MANIFEST_FILENAME);
-    let payload = serde_json::to_string_pretty(&entries)?;
-    std::fs::write(&manifest_path, payload)
-        .with_context(|| format!("Failed to write {}", manifest_path.display()))?;
-    Ok(())
-}
-
-/// Build the harness firmware required by a radio HIL test (if not already
-/// built) and return the resulting ELF path. Results are cached in
-/// `built_harness` so repeated tests sharing a harness only build it once.
-fn build_radio_harness(
-    workspace: &Path,
-    package_path: &Path,
-    chip: Chip,
-    target: &str,
-    toolchain: Option<&str>,
-    timings: bool,
-    harness: &firmware::Metadata,
-    built_harness: &mut HashMap<String, PathBuf>,
-) -> Result<PathBuf> {
-    let artifact_name = harness.output_file_name();
-    if let Some(path) = built_harness.get(&artifact_name) {
-        return Ok(path.clone());
-    }
-
-    let builder = crate::generate_build_command(
-        package_path,
-        chip,
-        target,
-        harness,
-        CargoAction::Build(None),
-        false,
-        toolchain,
-        timings,
-        &[],
-    )?;
-    let built = CargoCommandBatcher::build_one_for_cargo(&builder);
-    log::info!(
-        "Building harness firmware: cargo {}",
-        built.command.join(" ")
-    );
-    built
-        .run(false)
-        .with_context(|| format!("Failed to build harness firmware '{artifact_name}'"))?;
-
-    let path = crate::radio_hil_runner::resolve_release_binary(workspace, target, &artifact_name);
-    if !path.exists() {
-        bail!(
-            "Harness firmware '{artifact_name}' built successfully but binary not found at {}",
-            path.display()
-        );
-    }
-
-    built_harness.insert(artifact_name, path.clone());
-    Ok(path)
-}
-
-fn resolve_radio_module_alias(
-    all_tests: &[firmware::Metadata],
-    chip: Chip,
-    selected: &str,
-) -> Option<firmware::Metadata> {
-    let module_file = format!("{selected}.rs");
-    let module_decl = format!("mod {selected};");
-
-    all_tests
-        .iter()
-        .find(|test| {
-            test.supports_chip(chip)
-                && test.harness_firmware().is_none()
-                && std::fs::read_to_string(test.example_path()).is_ok_and(|source| {
-                    source.contains(&module_file) || source.contains(&module_decl)
-                })
-        })
-        .cloned()
-}
-
-fn is_support_firmware(meta: &firmware::Metadata) -> bool {
-    meta.is_support_firmware()
-}
-
-fn move_artifacts(chip: Chip, action: &CargoAction) {
+pub(crate) fn move_artifacts(chip: Chip, action: &CargoAction) {
     if let CargoAction::Build(Some(out_dir)) = action {
         // Move binaries
         let from = out_dir.join("tmp");
