+++ xtask/src/main.rs
@@ -1,22 +1,8 @@
-use std::{
-    path::{Path, PathBuf},
-    process::Command,
-    time::Instant,
-};
-
-use anyhow::{Context, Result, bail};
+use anyhow::{Context, Result};
 use clap::{Args, Parser};
 use esp_devtool as xtask;
-use object::{Object, ObjectSymbol, SymbolKind, read::archive::ArchiveFile};
-use rustc_demangle::try_demangle;
 use strum::IntoEnumIterator;
-use xtask::{
-    Package,
-    cargo::{CargoAction, CargoArgsBuilder, CargoCommandBatcher},
-    commands::*,
-    metadata::{Chip, Config},
-    resolve::Verb,
-};
+use xtask::{Package, commands::*, resolve::Verb};
 
 // ----------------------------------------------------------------------------
 // Command-line Interface
@@ -186,742 +172,5 @@ fn main() -> Result<()> {
             .enable_all()
             .build()?
             .block_on(xtask::commands::mcp::run_mcp_server()),
-    }
-}
-
-// ----------------------------------------------------------------------------
-// Subcommands
-
-fn fmt_packages(workspace: &Path, args: FmtPackagesArgs) -> Result<()> {
-    let mut packages = args.packages;
-    packages.sort();
-
-    for package in packages {
-        xtask::format_package(workspace, package, args.check, None)?;
-    }
-
-    // Keep formatting "helper" crates that are intentionally not in
-    // the Package enum.
-    for package_path in ["xtask", "xtask-mcp-macros"] {
-        log::info!("Formatting package: {}", package_path);
-        xtask::format_package_path(workspace, &workspace.join(package_path), args.check, None)?;
-    }
-
-    // format ymls in .github/
-    xtask::format_yml(args.check, "./.github")?;
-
-    Ok(())
-}
-
-fn clean(workspace: &Path, args: CleanArgs) -> Result<()> {
-    let mut packages = args.packages;
-    packages.sort();
-
-    for package in packages {
-        let path = workspace.join(package.directory());
-        for dir in walkdir::WalkDir::new(path) {
-            if let Ok(dir) = dir
-                && let path = dir.path()
-                && path.join("Cargo.toml").exists()
-            {
-                log::info!("Cleaning folder: {}", path.display());
-                let cargo_args = CargoArgsBuilder::default()
-                    .subcommand("clean")
-                    .arg("--target-dir")
-                    .arg(path.join("target").display().to_string())
-                    .build();
-
-                xtask::cargo::run(&cargo_args, path).with_context(|| {
-                    format!(
-                        "Failed to run `cargo run` with {cargo_args:?} in {}",
-                        path.display()
-                    )
-                })?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn check_packages(workspace: &Path, args: CheckPackagesArgs) -> Result<()> {
-    log::debug!("Checking packages: {:?}", args.packages);
-    let mut packages = args.packages;
-    packages.sort();
-
-    let mut commands = CargoCommandBatcher::new();
-
-    for package in packages.iter().filter(|p| p.is_published()) {
-        // Unfortunately each package has its own unique requirements for
-        // building, so we need to handle each individually (though there
-        // is *some* overlap)
-        for chip in &args.chips {
-            log::debug!("  for chip: {}", chip);
-            let device = Config::for_chip(chip);
-
-            if let Err(e) = package.validate_package_chip(chip) {
-                log::warn!("{e}. Skipping");
-                continue;
-            }
-
-            for mut check_config in package.check_config_rules(device) {
-                if package.has_chip_features() {
-                    check_config.features.push(device.name.clone());
-                }
-
-                commands.push(build_check_package_command(
-                    workspace,
-                    *package,
-                    chip,
-                    &["--no-default-features"],
-                    &check_config,
-                    args.toolchain.as_deref(),
-                )?);
-            }
-        }
-    }
-
-    for c in commands.build(false) {
-        println!(
-            "Command: cargo {}",
-            c.command.join(" ").replace("---", "\n    ---")
-        );
-        c.run(false)?;
-    }
-
-    Ok(())
-}
-
-fn build_check_package_command(
-    workspace: &Path,
-    package: Package,
-    chip: &Chip,
-    args: &[&str],
-    check_config: &xtask::CheckConfig,
-    mut toolchain: Option<&str>,
-) -> Result<CargoArgsBuilder> {
-    log::info!(
-        "Checking package: {} ({}, features: {:?}, env: {:?})",
-        package,
-        chip,
-        check_config.features,
-        check_config.env
-    );
-
-    let path = workspace.join(package.directory());
-    let features = &check_config.features;
-
-    let mut builder = CargoArgsBuilder::default()
-        .subcommand("check")
-        .manifest_path(path.join("Cargo.toml"));
-
-    if !package.build_on_host(features) {
-        if chip.is_xtensa() {
-            // In case the user doesn't specify a toolchain, make sure we use +esp
-            toolchain.get_or_insert("esp");
-        }
-        builder = builder.target(package.target_triple(chip)?);
-    }
-
-    if let Some(toolchain) = toolchain {
-        if !package.build_on_host(features) && toolchain.starts_with("esp") {
-            builder = builder.config("-Zbuild-std=core,alloc");
-        }
-        builder = builder.toolchain(toolchain);
-    }
-
-    builder = builder.args(args);
-
-    if !features.is_empty() {
-        builder = builder.arg(format!("--features={}", features.join(",")));
-    }
-
-    for (key, value) in &check_config.env {
-        builder.add_env_var(key, value);
-    }
-    // TODO: these should come from the outside
-    builder.add_env_var("CI", "1");
-    builder.add_env_var("DEFMT_LOG", "trace");
-    builder.add_env_var("ESP_LOG", "trace");
-
-    Ok(builder)
-}
-
-fn lint_packages(workspace: &Path, args: LintPackagesArgs) -> Result<()> {
-    log::debug!("Linting packages: {:?}", args.packages);
-    let mut packages = args.packages;
-    packages.sort();
-
-    for package in packages.iter().filter(|p| p.is_published()) {
-        // Unfortunately each package has its own unique requirements for
-        // building, so we need to handle each individually (though there
-        // is *some* overlap)
-        for chip in &args.chips {
-            log::debug!("  for chip: {}", chip);
-            let device = Config::for_chip(chip);
-
-            if let Err(e) = package.validate_package_chip(chip) {
-                log::warn!("{e}. Skipping");
-                continue;
-            }
-
-            for mut check_config in package.lint_config_rules(device) {
-                if package.has_chip_features() {
-                    check_config.features.push(device.name.clone());
-                }
-
-                lint_package(
-                    workspace,
-                    *package,
-                    chip,
-                    &["--no-default-features"],
-                    &check_config,
-                    args.fix,
-                    args.toolchain.as_deref(),
-                )?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn lint_package(
-    workspace: &Path,
-    package: Package,
-    chip: &Chip,
-    args: &[&str],
-    check_config: &xtask::CheckConfig,
-    fix: bool,
-    mut toolchain: Option<&str>,
-) -> Result<()> {
-    log::info!(
-        "Linting package: {} ({}, features: {:?}, env: {:?})",
-        package,
-        chip,
-        check_config.features,
-        check_config.env
-    );
-
-    let path = workspace.join(package.directory());
-    let features = &check_config.features;
-
-    let mut builder = CargoArgsBuilder::default()
-        .subcommand("clippy")
-        .manifest_path(path.join("Cargo.toml"));
-
-    if !package.build_on_host(features) {
-        if chip.is_xtensa() {
-            // In case the user doesn't specify a toolchain, make sure we use +esp
-            toolchain.get_or_insert("esp");
-        }
-        builder = builder.target(package.target_triple(chip)?);
-    }
-
-    if let Some(toolchain) = toolchain {
-        if !package.build_on_host(features) && toolchain.starts_with("esp") {
-            builder = builder.config("-Zbuild-std=core,alloc");
-        }
-        builder = builder.toolchain(toolchain);
-    }
-
-    for arg in args {
-        builder = builder.arg(arg.to_string());
-    }
-
-    if !features.is_empty() {
-        builder = builder.arg(format!("--features={}", features.join(",")));
-    }
-
-    let mut builder = if fix {
-        builder.arg("--fix").arg("--lib").arg("--allow-dirty")
-    } else {
-        builder.arg("--").arg("-D").arg("warnings").arg("--no-deps")
-    };
-
-    for (key, value) in &check_config.env {
-        builder.add_env_var(key, value);
-    }
-    builder.add_env_var("CI", "1");
-    builder.add_env_var("DEFMT_LOG", "trace");
-
-    let command = CargoCommandBatcher::build_one_for_cargo(&builder);
-    xtask::cargo::run_with_env(&command.command, &path, command.env_vars, false)
-        .with_context(|| format!("Failed to run `cargo clippy` in {}", path.display()))?;
-
-    Ok(())
-}
-
-struct Runner {
-    failed: Vec<&'static str>,
-    started_at: Instant,
-    skip_steps: Vec<String>,
-    run_steps: Vec<String>,
-    steps_executed: usize,
-}
-
-impl Runner {
-    fn new(options: &CiArgs) -> Self {
-        Self {
-            failed: Vec::new(),
-            started_at: Instant::now(),
-            skip_steps: {
-                let mut skip = vec![];
-                if options.no_lint {
-                    skip.push(String::from("lint"));
-                }
-                if options.no_docs {
-                    skip.push(String::from("docs"));
-                    skip.push(String::from("doc-tests"));
-                }
-                if options.no_check_crates {
-                    skip.push(String::from("check"));
-                }
-                skip
-            },
-            run_steps: options.steps.clone(),
-            steps_executed: 0,
-        }
-    }
-
-    fn run(&mut self, id: &str, group: &'static str, op: impl FnOnce() -> Result<()>) {
-        if self.skip_steps.iter().any(|s| s == id) {
-            log::debug!("{group} skipped by user request");
-            return;
-        }
-        if !self.run_steps.is_empty() && !self.run_steps.iter().any(|s| s == id) {
-            log::debug!("{group} skipped by user request");
-            return;
-        }
-
-        self.steps_executed += 1;
-
-        // Output grouped logs
-        // https://docs.github.com/en/actions/reference/workflows-and-actions/workflow-commands#grouping-log-lines
-        println!("::group::{group}");
-        if let Err(e) = op() {
-            log::error!("{group} failed: {e:?}");
-            self.failed.push(group);
-        } else {
-            log::debug!("{group} succeeded");
-        }
-        println!("::endgroup::");
-    }
-
-    fn finish(self) -> Result<()> {
-        fn write_summary(message: &str) {
-            if let Some(summary_file) = std::env::var_os("GITHUB_STEP_SUMMARY") {
-                std::fs::write(summary_file, message).unwrap();
-            }
-        }
-
-        let expected_to_run = self
-            .run_steps
-            .iter()
-            .filter(|s| !self.skip_steps.contains(s))
-            .cloned()
-            .collect::<Vec<_>>();
-        if self.steps_executed == 0 && !expected_to_run.is_empty() {
-            bail!(
-                "The following steps were requested but not executed: {}. Perhaps they contain typos?",
-                expected_to_run.join(", ")
-            );
-        }
-
-        log::info!("CI checks completed in {:?}", self.started_at.elapsed());
-
-        if !self.failed.is_empty() {
-            let mut summary = String::new();
-            summary.push_str("# Summary of failed CI checks\n");
-            for failed_check in self.failed {
-                summary.push_str(&format!("* {failed_check}\n"));
-            }
-            println!("{summary}");
-            write_summary(&summary);
-            bail!("CI checks failed");
-        }
-
-        Ok(())
-    }
-}
-
-fn run_ci_checks(workspace: &Path, args: CiArgs) -> Result<()> {
-    log::info!("Running CI checks for chip: {}", args.chip);
-    println!("::add-matcher::.github/rust-matchers.json");
-
-    let run_locally = !std::env::var("CI").is_ok();
-
-    let mut runner = Runner::new(&args);
-
-    if !run_locally && args.steps.is_empty() {
-        runner.skip_steps.push(String::from("tests"));
-    }
-
-    unsafe {
-        std::env::set_var("CI", "true");
-    }
-
-    runner.run("check", "Check crates", || {
-        check_packages(
-            workspace,
-            CheckPackagesArgs {
-                packages: Package::iter().collect(),
-                chips: vec![args.chip],
-                toolchain: args.toolchain.clone(),
-            },
-        )
-    });
-
-    runner.run("lint", "Lint", || {
-        lint_packages(
-            workspace,
-            LintPackagesArgs {
-                packages: Package::iter().collect(),
-                chips: vec![args.chip],
-                fix: false,
-                toolchain: args.toolchain.clone(),
-            },
-        )
-    });
-
-    runner.run("doc-tests", "Run Doc Test", || {
-        run_doc_tests(
-            workspace,
-            DocTestArgs {
-                packages: Package::iter().collect(),
-                chip: args.chip,
-            },
-        )
-    });
-
-    runner.run("docs", "Build Docs", || {
-        build_documentation(
-            workspace,
-            BuildDocumentationArgs {
-                packages: vec![Package::EspHal, Package::EspRadio, Package::EspRtos],
-                chips: vec![args.chip],
-                ..Default::default()
-            },
-        )
-    });
-
-    // for chips with esp-lp-hal: Build all supported examples for the low-power
-    // core first
-    if args.chip.has_lp_core() {
-        // Build prerequisite examples (esp-lp-hal)
-        // `examples` copies the examples to a folder with the chip name as the last
-        // path element then we copy it to the place where the HP core example
-        // expects it
-        runner.run("lp-examples", "Build LP-HAL Examples", || {
-            // The LP examples aren't really that demanding, but they need to be at a certain place.
-            // Instead of trying to figure out where the results are, let's just make sure the
-            // target folder is set up as expected.
-            let original_target_dir = std::env::var("CARGO_TARGET_DIR");
-
-            unsafe {
-                std::env::set_var(
-                    "CARGO_TARGET_DIR",
-                    workspace.join("esp-lp-hal").join("target"),
-                );
-            }
-            let result = examples(
-                workspace,
-                Package::EspLpHal,
-                args.chip,
-                Some("all"),
-                CargoAction::Build(None),
-                false,
-                args.toolchain.as_deref(),
-                false,
-            );
-
-            // Still need to rename examples to remove the fingerprint off of their names:
-            let dir = workspace
-                .join("esp-lp-hal")
-                .join("target")
-                .join(Package::EspLpHal.target_triple(&args.chip)?)
-                .join("release")
-                .join("examples");
-            let examples = dir
-                .read_dir()
-                .with_context(|| format!("Failed to read examples directory: {}", dir.display()))?;
-            for example in examples {
-                let example = example.context("Failed to read example")?;
-                let example_path = example.path();
-                if example
-                    .file_type()
-                    .with_context(|| {
-                        format!(
-                            "Failed to get file type for example: {}",
-                            example_path.display()
-                        )
-                    })?
-                    .is_file()
-                    && example_path.extension().is_none()
-                {
-                    let example_name = example.file_name().to_string_lossy().to_string();
-                    let Some((without_fingerprint, fingerprint)) = example_name.rsplit_once('-')
-                    else {
-                        // Skip non-fingerprinted files to avoid self-copying
-                        // artifacts (for example "blinky" -> "blinky").
-                        continue;
-                    };
-                    // Cargo fingerprints are hexadecimal, so only rewrite files
-                    // that actually look like fingerprinted artifacts.
-                    if !fingerprint.chars().all(|c| c.is_ascii_hexdigit()) {
-                        continue;
-                    }
-
-                    let dst = dir.join(without_fingerprint);
-
-                    if dst.exists() {
-                        std::fs::remove_file(&dst).with_context(|| {
-                            format!(
-                                "Failed to remove destination file {} before copying",
-                                dst.display()
-                            )
-                        })?;
-                    }
-
-                    log::debug!("Copying {} to {}", example_path.display(), dst.display());
-
-                    // Copy so we don't trigger a rebuild unnecessarily by deleting the original
-                    std::fs::copy(&example_path, &dst).with_context(|| {
-                        format!(
-                            "Failed to copy example: {} to {}",
-                            example_path.display(),
-                            dst.display()
-                        )
-                    })?;
-                    // Check that the destination file matches the source file
-                    let dst_contents = std::fs::read(&dst).with_context(|| {
-                        format!("Failed to read destination file: {}", dst.display())
-                    })?;
-                    let src_contents = std::fs::read(&example_path).with_context(|| {
-                        format!("Failed to read source file: {}", example_path.display())
-                    })?;
-                    if dst_contents != src_contents {
-                        return Err(anyhow::anyhow!(
-                            "Destination file {} does not match source file {}",
-                            dst.display(),
-                            example_path.display()
-                        ));
-                    }
-                }
-            }
-
-            // Restore the original target directory
-            unsafe {
-                if let Ok(target) = original_target_dir {
-                    std::env::set_var("CARGO_TARGET_DIR", target);
-                } else {
-                    std::env::remove_var("CARGO_TARGET_DIR");
-                }
-            }
-
-            result
-        });
-
-        // Check documentation. Reuse the "docs" ID, as the docs build should include all applicable
-        // packages.
-        runner.run("docs", "Build LP-HAL docs", || {
-            build_documentation(
-                workspace,
-                BuildDocumentationArgs {
-                    packages: vec![Package::EspLpHal],
-                    chips: vec![args.chip],
-                    ..Default::default()
-                },
-            )
-        });
-    }
-
-    runner.run("examples", "Build examples", || {
-        // The `ota_example` expects a file named `target/ota_image` - it doesn't
-        // care about the contents however
-        std::fs::create_dir_all("./target")
-            .with_context(|| "Failed to create `./target`".to_string())?;
-        std::fs::write("./target/ota_image", "DUMMY").with_context(|| {
-            "Failed to create a dummy file required by ota example!".to_string()
-        })?;
-
-        examples(
-            workspace,
-            Package::Examples,
-            args.chip,
-            Some("all"),
-            CargoAction::Build(None),
-            true,
-            args.toolchain.as_deref(),
-            false,
-        )
-    });
-
-    runner.run("qa-test", "Build qa-test", || {
-        examples(
-            workspace,
-            Package::QaTest,
-            args.chip,
-            Some("all"),
-            CargoAction::Build(None),
-            true,
-            args.toolchain.as_deref(),
-            false,
-        )
-    });
-
-    // Try-build tests
-    runner.run("tests", "Build tests", || {
-        let target_path = workspace.join("target");
-
-        tests(
-            workspace,
-            Package::HilTest,
-            args.chip,
-            None,
-            CargoAction::Build(Some(target_path.join("tests"))),
-            1,
-            None,
-            false,
-        )?;
-
-        tests(
-            workspace,
-            Package::HilTestRadio,
-            args.chip,
-            None,
-            CargoAction::Build(Some(target_path.join("tests"))),
-            1,
-            None,
-            false,
-        )
-    });
-
-    runner.finish()
-}
-
-fn host_tests(workspace: &Path, args: HostTestsArgs) -> Result<()> {
-    let mut packages = args.packages;
-    packages.sort();
-
-    for package in packages {
-        log::debug!("Running host-tests for package: {}", package);
-        if package.has_host_tests(workspace) {
-            xtask::run_host_tests(workspace, package)?;
-        }
-    }
-
-    Ok(())
-}
-
-/// Build the given package for the given chip and return its `.rlib path` with `esp` toolchain.
-fn build_rlib(package: &str, chip: &str, target: &str) -> Result<PathBuf> {
-    let workspace = std::env::current_dir().with_context(|| "Failed to get the current dir!")?;
-    Command::new("cargo")
-        .args([
-            "+esp",
-            "build",
-            "--no-default-features",
-            "--features",
-            chip,
-            "--target",
-            target,
-            "-Zbuild-std=core",
-        ])
-        .current_dir(workspace.join(package))
-        .status()
-        .context("Failed to run cargo build")?;
-
-    let lib_name = format!("lib{}.rlib", package.replace('-', "_"));
-    let rlib_path = Path::new("target")
-        .join(target)
-        .join("debug")
-        .join(&lib_name);
-
-    if !rlib_path.exists() {
-        anyhow::bail!("Expected rlib not found: {}", rlib_path.display());
-    }
-
-    Ok(rlib_path)
-}
-
-/// Check global symbols in the compiled rlib of the specified packages for the
-/// specified chips. Reports any unmangled global symbols that may pollute the
-/// global namespace.
-fn check_global_symbols(chips: &[Chip]) -> Result<()> {
-    let mut total_problematic = vec![];
-
-    let package = Package::EspHal; // Only esp-hal for now
-
-    for chip in chips {
-        let target = package.target_triple(chip)?;
-
-        let rlib_path = match build_rlib(package.as_ref(), chip.as_ref(), &target) {
-            Ok(path) => path,
-            Err(e) => {
-                println!(
-                    "Failed to build/find rlib for {} on {}: {}",
-                    package, chip, e
-                );
-                continue;
-            }
-        };
-
-        log::info!("Checking global symbols for {} on {}:\n", package, chip);
-
-        let data = std::fs::read(&rlib_path)
-            .with_context(|| format!("Failed to read {}!", rlib_path.display()))?;
-        let archive = ArchiveFile::parse(data.as_slice())
-            .with_context(|| "Failed to create archive!".to_string())?;
-
-        let mut problematic_symbols: Vec<(String, SymbolKind, usize)> = Vec::new();
-
-        for member in archive.members().flatten() {
-            let obj = object::File::parse(
-                member
-                    .data(data.as_slice())
-                    .with_context(|| "Failed to parse archive!")?,
-            )?;
-
-            for symbol in obj.symbols().filter(|s| s.is_global() && s.is_definition()) {
-                if let Ok(name) = symbol.name()
-                    && try_demangle(name).is_err()
-                {
-                    let section = symbol.section_index().map(|i| i.0).unwrap_or(0);
-                    problematic_symbols.push((name.to_string(), symbol.kind(), section));
-                }
-            }
-        }
-
-        if problematic_symbols.is_empty() {
-            println!("All global symbols are properly mangled Rust symbols\n");
-        } else {
-            println!(
-                "Found {} potentially problematic global symbols:",
-                problematic_symbols.len(),
-            );
-
-            for (name, kind, _) in &problematic_symbols {
-                println!("{:?} {}", kind, name);
-            }
-
-            total_problematic.extend(
-                problematic_symbols
-                    .into_iter()
-                    .map(|(name, kind, _)| (chip, name, kind)),
-            );
-        }
-    }
-
-    if !total_problematic.is_empty() {
-        for (chip, name, kind) in total_problematic.iter() {
-            println!("{}: {} ({:?})", chip, name, kind);
-        }
-        Err(anyhow::anyhow!(
-            "Found {count} unmangled global symbols across all packages/chips",
-            count = total_problematic.len()
-        ))
-    } else {
-        Ok(())
     }
 }
